:recycle: Refactor: 수익률 계산 수식 수정

+++ src/main/java/lotto/domain/ProfitCalculateStrategy.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
 public interface ProfitCalculateStrategy {
-    
-    Integer calculateProfitRate(int totalPrice, int money);
+
+    double calculateProfitRate(int totalPrice, int money);
 }
